+++ quasitools/commands/cmd_hydra.py
@@ -1,7 +1,7 @@
 """
-Copyright Government of Canada 2017
+Copyright Government of Canada 2017 - 2018
 
-Written by: Camy Tran, National Microbiology Laboratory,
+Written by: Camy Tran and Matthew Fogel, National Microbiology Laboratory,
             Public Health Agency of Canada
 
 Licensed under the Apache License, Version 2.0 (the "License"); you may not use
@@ -20,6 +20,21 @@ import os
 from collections import defaultdict
 import click
 from quasitools.patient_analyzer import PatientAnalyzer
+
+# GLOBALS
+
+from quasitools.quality_control import TRIMMING
+from quasitools.quality_control import MASKING
+from quasitools.quality_control import MIN_READ_QUAL
+from quasitools.quality_control import LENGTH_CUTOFF
+from quasitools.quality_control import MEDIAN_CUTOFF
+from quasitools.quality_control import MEAN_CUTOFF
+from quasitools.quality_control import NS
+from quasitools.patient_analyzer import ERROR_RATE
+from quasitools.patient_analyzer import MIN_VARIANT_QUAL
+from quasitools.patient_analyzer import MIN_DP
+from quasitools.patient_analyzer import MIN_AC
+from quasitools.patient_analyzer import MIN_FREQ
 
 BASE_PATH = os.path.abspath(os.path.join(os.path.abspath(__file__),
                                          os.pardir, os.pardir, "data"))
@@ -42,43 +57,67 @@ MUTATION_DB = os.path.join(BASE_PATH, "mutation_db.tsv")
               default=MUTATION_DB)
 @click.option('-rt', '--reporting_threshold', default=1,
               type=click.IntRange(1, 100, clamp=True),
-              help='minimum mutation frequency percent to report.')
+              help='Minimum mutation frequency percent to report.')
 @click.option('-gc', '--generate_consensus',
               help='Generate a mixed base consensus sequence.', is_flag=True)
 @click.option('-cp', '--consensus_pct', default=20,
               type=click.IntRange(1, 100, clamp=True),
-              help='minimum percentage a base needs to be incorporated'
+              help='Minimum percentage a base needs to be incorporated '
               'into the consensus sequence.')
 @click.option('-q', '--quiet', is_flag=True,
-              help='suppress all normal output')
+              help='Suppress all normal output.')
+@click.option('-tr', '--trim_reads', is_flag=True,
+              help='Iteratively trim reads based on filter values if enabled. '
+                   'Remove reads which do not meet filter values if disabled.')
+@click.option('-mr', '--mask_reads', is_flag=True,
+              help='Mask low coverage regions in reads if below minimum read'
+              ' quality. This option and --ns cannot be both enabled '
+              'simultaneously.')
+@click.option('-rq', '--min_read_qual', default=30, help='Minimum quality for '
+              'a position in a read to be masked.')
 @click.option('-lc', '--length_cutoff', default=100,
-              help='reads which fall short of the specified length '
+              help='Reads which fall short of the specified length '
                    'will be filtered out.')
 @click.option('-sc', '--score_cutoff', default=30,
-              help='reads whose average quality score is less than the '
-                   'specified score will be filtered out.')
-@click.option('-n', '--ns', is_flag=True, help='flag to enable the '
-              'filtering of n\'s')
+              help='Reads that have a median or mean quality score (depending'
+                   ' on the score type specified) less than the score cutoff '
+                   'value will be filtered out.')
+@click.option('-me/-mn', '--median/--mean', 'score_type',
+              default=True,
+              help='Use either median score (default) or mean score for the '
+              'score cutoff value.')
+@click.option('-n', '--ns', is_flag=True, help='Flag to enable the '
+              'filtering of n\'s.  This option and --mask_reads cannot be both'
+              ' enabled simultaneously.')
 @click.option('-e', '--error_rate', default=0.0021,
-              help='error rate for the sequencing platform.')
-@click.option('-mq', '--min_qual', default=30, help='minimum quality for '
-              'variant to be considered later on in the pipeline.')
+              help='Error rate for the sequencing platform.')
+@click.option('-vq', '--min_variant_qual', default=30, help='Minimum quality '
+              'for variant to be considered later on in the pipeline.')
 @click.option('-md', '--min_dp', default=100,
-              help='minimum required read depth for.')
+              help='Minimum required read depth for variant to be considered'
+              ' later on in the pipeline.')
 @click.option('-ma', '--min_ac', default=5,
-              help='the minimum required allele count.')
+              help='The minimum required allele count for variant to be '
+              'considered later on in the pipeline.')
 @click.option('-mf', '--min_freq', default=0.01,
-              help='the minimum required frequency.')
+              help='The minimum required frequency for mutation to be '
+              'considered in drug resistance report.')
 @click.option('-i', '--id',
-              help='specify FASTA sequence identifier to be used in the '
+              help='Specify FASTA sequence identifier to be used in the '
               'consensus report.')
 @click.pass_context
 def cli(ctx, output_dir, forward, reverse, mutation_db, reporting_threshold,
-        generate_consensus, consensus_pct, quiet, length_cutoff,
-        score_cutoff, ns, error_rate, min_qual, min_dp, min_ac, min_freq, id):
+        generate_consensus, consensus_pct, quiet, trim_reads, mask_reads,
+        min_read_qual, length_cutoff, score_cutoff, score_type, ns, error_rate,
+        min_variant_qual, min_dp, min_ac, min_freq, id):
 
-    os.mkdir(output_dir)
+    if not os.path.isdir(output_dir):
+        os.mkdir(output_dir)
     reads = forward
+
+    if mask_reads and ns:
+        raise click.UsageError("--ns and --mask_reads cannot be enabled" +
+                               " simultaneously.")
 
     # The fasta_id is used as the sequence id in the consensus report
     # and as the RG-ID in the bt2-generated bam file.
@@ -98,6 +137,36 @@ def cli(ctx, output_dir, forward, reverse, mutation_db, reporting_threshold,
         if not id:
             fasta_id += ("_%s" % os.path.basename(reverse).split('.')[0])
 
+    quality_filters = defaultdict(dict)
+
+    if trim_reads:
+        quality_filters[TRIMMING] = True
+
+    if mask_reads:
+        quality_filters[MASKING] = True
+
+    quality_filters[LENGTH_CUTOFF] = length_cutoff
+
+    if score_type == "median":
+        quality_filters[MEDIAN_CUTOFF] = score_cutoff
+    elif score_type == "mean":
+        quality_filters[MEAN_CUTOFF] = score_cutoff
+    # end if
+
+    if ns:
+        quality_filters[NS] = True
+    else:
+        quality_filters[NS] = False
+
+    quality_filters[MIN_READ_QUAL] = min_read_qual
+
+    variant_filters = defaultdict(dict)
+    variant_filters[ERROR_RATE] = error_rate
+    variant_filters[MIN_VARIANT_QUAL] = min_variant_qual
+    variant_filters[MIN_DP] = min_dp
+    variant_filters[MIN_AC] = min_ac
+    variant_filters[MIN_FREQ] = min_freq
+
     patient_analyzer = PatientAnalyzer(id=REFERENCE[REFERENCE.rfind('/')+1:],
                                        output_dir=output_dir,
                                        reads=reads, reference=REFERENCE,
@@ -105,22 +174,6 @@ def cli(ctx, output_dir, forward, reverse, mutation_db, reporting_threshold,
                                        mutation_db=mutation_db, quiet=quiet,
                                        consensus_pct=consensus_pct)
 
-    read_filters = defaultdict(dict)
-    read_filters["length_cutoff"] = length_cutoff
-    read_filters["score_cutoff"] = score_cutoff
-    if ns:
-        read_filters["ns"] = True
-    else:
-        read_filters["ns"] = False
-
-    patient_analyzer.filter_reads(read_filters)
-
-    variant_filters = defaultdict(dict)
-    variant_filters["error_rate"] = error_rate
-    variant_filters["min_qual"] = min_qual
-    variant_filters["min_dp"] = min_dp
-    variant_filters["min_ac"] = min_ac
-    variant_filters["min_freq"] = min_freq
-
+    patient_analyzer.filter_reads(quality_filters)
     patient_analyzer.analyze_reads(fasta_id, variant_filters,
                                    reporting_threshold, generate_consensus)